Split seamless config into separate file
